Normalización de clases y solución de errores

+++ src/main/java/org/nexus/proyecto_fn_nexus/dominio/exception/InvitadoNoExisteException.java
@@ -1,7 +1,7 @@
 package org.nexus.proyecto_fn_nexus.dominio.exception;
 
 public class InvitadoNoExisteException extends RuntimeException {
-    public InvitadoNoExisteException(Integer idInvitado) {
+    public InvitadoNoExisteException(Long idInvitado) {
         super("El invitado con ID: " + idInvitado + " no existe en el sistema");
     }
 }
