+++ src/main/java/org/nexus/proyecto_fn_nexus/dominio/exception/RecursoNoExisteException.java
@@ -1,7 +1,7 @@
 package org.nexus.proyecto_fn_nexus.dominio.exception;
 
 public class RecursoNoExisteException extends RuntimeException {
-    public RecursoNoExisteException(Integer idRecurso) {
+    public RecursoNoExisteException(Long idRecurso) {
         super("El recurso con codigo: " + idRecurso + " no existe");
     }
 }
