+++ src/main/java/org/nexus/proyecto_fn_nexus/persistence/crud/CrudRecursoEntity.java
@@ -3,7 +3,7 @@ package org.nexus.proyecto_fn_nexus.persistence.crud;
 import org.nexus.proyecto_fn_nexus.persistence.entity.RecursoEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CrudRecursoEntity extends CrudRepository<RecursoEntity, Integer> {
+public interface CrudRecursoEntity extends CrudRepository<RecursoEntity, Long> {
 
     RecursoEntity findFirstByNombreRecurso(String nombreRecurso);
 }
